Refactored MessageButtonGroup component outside Message page; Created unit tests for MessageButtonGroup
