+++ src/components/Messages/index.js
@@ -1,5 +1,6 @@
 import MessagePreview from './MessagePreview';
 import MessageFolder from './MessageFolder';
+import MessageButtonGroup from './MessageButtonGroup';
 
 /**
  * Components and functions related to Messages functionality within project PASS
@@ -7,4 +8,4 @@ import MessageFolder from './MessageFolder';
  * @namespace Messages
  */
 
-export { MessagePreview, MessageFolder };
+export { MessagePreview, MessageFolder, MessageButtonGroup };
